rename upload -> gcs-tools
